requirements.txt
fastapi==0.110.2
pydantic==2.7.1
uvicorn==0.15.0
aiofiles==23.2.1
PyJWT==2.8.0
pydantic[email]==2.7.1